PHPCS: fix up the code base [5] - multi-line function calls
Multi-line function calls need to have each argument on a new line.
In a next iteration of this principle, it is expected that a sniff will be introduced to ban multi-line function call arguments.

With this mind, this function call with multi-line parameters which was already causing errors to be thrown by PHPCS, has been fixed by moving multi-line function call arguments out of the function call and defining these as a variable before passing it to the function call.

The multiple nested function calls have also been unnested.